Adapt InitJira init step to API change in the InitIts init step
With change [1] in plugins/its-base support for configuring branches
for the Jira integration during init was added. Due to this change the
InitIts init step needs to be adapted.

In addition update the example of the plugin initialisation in the
documentation because with change [1] there is now an additional
question during init for configuring the branches.

[1] https://gerrit-review.googlesource.com/53690

Change-Id: Ifcfd87dd1918f4db8960271f9717d139160085b2
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>