Remove plugin name from build configuration
Removing the pluginName from the build configuration allows the Gerrit
plugin loader to dynamically set the plugin name to the same as the file
name (minus the extension). This makes it a little more flexible for the
end user.  In cases where the end users wants to change the plugin name
they only need to rename the plugin file instead of having to rebuild
the plugin.

This is a follow on to change Ib685a802 which documents a use case
for changing an its plugin file name.

Change-Id: I8a625b470a38757af490e2f48f251a48ac898c06