Use REST APIs instead of SOAP APIs for Gerrit 2.11
The plugin does not work currently for Gerrit 2.11 and Jira 7.
Jira 7 only supports REST API (the support of SOAP API was removed)
but the switch to REST API for the plugin was only developed
on the branch stable-2.12

This version cherry-picks what has been done on stable-2.12
(i.e. commit 894eae1a423fd510ac90dea108cb004768b1630b)
and update the imports so that they work for Gerrit 2.11
(its-base on stable-2.11 still uses "hooks" hierarchy
and not "its" hierarchy as folder tree).

Change-Id: I3faab757f5a15e3043cc92f89c52ba89d3411d78