Update patch set 2
Patch Set 2:

We are still on gerrit 2.11 because we use a git multisite solution that doesn't support gerrit 2.12 yet.

The point of this change was only to share what I did so that it will be ready to use if anyone ends up in the same situation.

(Patch set 2 only removes a patch file that was committed by mistake.)

Patch-set: 2