Replace Atlassian REST Library with simple REST calls
The Atlassian library is a huge library containing a lot of
functionality. It has dependencies to libraries that conflict
with Gerrit Core libraries. It also makes it impossible to
install the its-jira plugin multiple times.

This change makes the implementation a whole lot simpler,
and only uses the REST calls it actually needs.

It also solves the issue on Gerrit Core where an exception
is thrown every second in an endless loop which seems to pop
up more frequent on Jira 7.2.7. This exception comes out of
the Atlassian library:
[2017-03-16 21:49:18,799] [httpclient-io:thread-4] ERROR org.apache.http.nio.protocol.HttpAsyncRequestExecutor : http-outgoing-254 [CLOSED] HTTP protocol exception: null
java.nio.channels.ClosedChannelException
        at sun.nio.ch.SocketChannelImpl.ensureReadOpen(SocketChannelImpl.java:257)
        at sun.nio.ch.SocketChannelImpl.read(SocketChannelImpl.java:300)
        at org.apache.http.nio.reactor.ssl.SSLIOSession.receiveEncryptedData(SSLIOSession.java:340)
        at org.apache.http.nio.reactor.ssl.SSLIOSession.isAppInputReady(SSLIOSession.java:366)
        at org.apache.http.impl.nio.reactor.AbstractIODispatch.inputReady(AbstractIODispatch.java:118)
        at org.apache.http.impl.nio.reactor.BaseIOReactor.validate(BaseIOReactor.java:220)
        at org.apache.http.impl.nio.reactor.AbstractIOReactor.execute(AbstractIOReactor.java:284)
        at org.apache.http.impl.nio.reactor.BaseIOReactor.execute(BaseIOReactor.java:106)
        at org.apache.http.impl.nio.reactor.AbstractMultiworkerIOReactor$Worker.run(AbstractMultiworkerIOReactor.java:604)
        at java.lang.Thread.run(Thread.java:745)

Change-Id: I50f66e7e0166fd2c26bfbad1ad29c76088371da9
Signed-off-by: Remy Bohmer <linux@bohmer.net>