Replace Atlassian Library with vanilla HTTP calls
The Atlassian library is a huge jar containing a lot of
unneeded functionality. It has dependencies to libraries that conflict
with Gerrit Core libraries. It also makes it impossible to
install the its-jira plugin multiple times.

This change makes the implementation a whole lot simpler,
and uses ONLY the Jira REST calls strictly needed.

It also solves the issue on Gerrit Core where an exception
is thrown every second in an endless loop which seems to pop
up more frequent on Jira 7.2.7. 
This exception comes out of the Atlassian library:
 [CLOSED] HTTP protocol exception: null
java.nio.channels.ClosedChannelException

Change-Id: I50f66e7e0166fd2c26bfbad1ad29c76088371da9
Signed-off-by: Remy Bohmer <linux@bohmer.net>
Signed-off-by: Luca Milanesio <luca.milanesio@gmail.com>