Support multiple Jira instances
Currently, its-jira only supports connecting to a single Jira server. In
corporate environments, it is not unusual to have multiple Jira servers
and it is a common requirement to integrate Gerrit with those.

This change enable its-jira to integrate with multiple Jira servers at
the Gerrit project level, i.e., a Gerrit project can be associated with
a particular Jira instance. This is done by specifying the Jira server
URL, username and password in the project configuration using the GUI
controls on the project's General page. The plugin adds a "commentlink"
section accordingly.

A typical Jira server configuration in project.config will look like:

   [plugin "its-jira"]
     enabled = true
     url = http://localhost:8075/
     username = admin
     password = admin

   [commentlink "its-jira"]
     match = ([A-Z]+-[0-9]+)
     link = http://localhost:8075/browse/$1

In case its-jira plugin is enabled for a project but no Jira server is
configured for this project, the default configuration will be the one
defined in gerrit.config. If no Jira server information is defined in
gerrit.config, an error is logged and the Jira integration is disabled
for the project.

Change-Id: I88f7298055ec1a136348758d4ec6fab67668e1ca