Support multiple Jira instances
Currently, its-jira only supports connecting to an unique Jira server.
In corporate environments is not unusual to have multiple Jira instances
and it is a common requirement to integrate Gerrit with those.

This change enable its-jira to integrate with multiple Jira servers at
the Gerrit project level, i.e., a Gerrit project or group of projects
can be associated with a Jira instance.

Users can configure a Jira server for a project using GUI controls on
the project's General page, specifying the Jira server url, username and
password. The plugin takes care of creating a "commentlink" section in
the project.config file accordingly. Is worth noting that the password
will be encrypted right after being typed in the UI.

A typical Jira server configuration in project.config will look like:

[plugin "its-jira"]
  enabled = true
  url = http://localhost:8075/
  username = admin
  password = admin
[commentlink "its-jira"]
  match = ([A-Z]+-[0-9]+)
  link = http://localhost:8075/browse/$1

If its-jira plugin is enabled for a project and no Jira server is
configured for this project, the default configuration will be the one
defined in gerrit.config.

In order to get the project configuration, this plugin needs to access
the name of the project currently being processed; given this info
was only available to its-base, the project name is sent to its-jira
through the ItsServer interface and the respective Jira server data is
tunnelled back to its-base so it can be passed as a parameter to any
action on the ITS system.

This change depends on an open change in its-base[1] (patch set 14)
which introduce the aforementioned ItsServer interface.

[1] https://gerrit-review.googlesource.com/c/plugins/its-base/+/148715/14

Change-Id: I88f7298055ec1a136348758d4ec6fab67668e1ca