Support multiple Jira instances
Currently, its-jira only supports connecting to an unique Jira server.
In corporate environments, it is not unusual to have multiple Jira instances
and is a common requirement to integrate Gerrit with those.

This change enable its-jira to integrate with multiple Jira servers at
the Gerrit project level i.e., a Gerrit project or group of projects
can be associated with a Jira instance.

Users can configure a Jira server for a project using GUI controls on
the project's General page, specifying the Jira server url, username and
password. The plugin takes care of creating a "commentlink" section in
the project.config file accordingly. It is worth noting that the password
will be encrypted right after being typed in the UI.

A typical Jira server configuration in project.config will look like:

[plugin "its-jira"]
  enabled = true
  url = http://localhost:8075/
  username = admin
  password = admin
[commentlink "its-jira"]
  match = ([A-Z]+-[0-9]+)
  link = http://localhost:8075/browse/$1

If its-jira plugin is enabled for a project and no Jira server is
configured for this project, the default configuration will be the one
defined in gerrit.config.

The  plugin its-jira makes use of the ItsFacadeFactory to create in
order to get the name of the project currently using its-jira.

This change depends on an open change in its-base[1]
which introduce the aforementioned ItsFacadeFactory interface and on
gerrit core change[2] which introduced the encryption API.

[1] https://gerrit-review.googlesource.com/c/plugins/its-base/+/148715/14
[2] https://gerrit-review.googlesource.com/c/gerrit/+/177390

Change-Id: I88f7298055ec1a136348758d4ec6fab67668e1ca