Sonar: Constructors should not be used to instantiate "String"
Constructors for Strings and the objects used to wrap primitives should
never be used. Doing so is less clear and uses more memory than simply
using the desired value in the case of strings, and using valueOf
for everything else.

Further, these constructors are deprecated in Java 9, which is an
indication that they will eventually be removed from the language
altogether.

Change-Id: Idabc1f446a836ba6d77b164f55bbfe6ac4e2ca9d