Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> I believe .settings should have never been checked in as it is generated through Bazel. Shall we just remove the directory altogether and put into the .gitignore?

It is not generated via bazel.  This is generated by Eclipse, and we need it to make Eclipse build the code with the correct Java version.

We do the same in core Gerrit and in other plugins.

Patch-set: 1
Reviewer: David Pursehouse <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>