Extract Jira configuration
Instead of reading the plugin configuration in the facade, extract the
handling of the configuration to its own class (with its corresponding
tests).

This change is done, first, to keep the separation of concerns between
classes and, two, it is required to implement multi Jira server
capabilities which is done in a follow up commit.

Change-Id: I63a0b997d241fb520715c29bc8bd46e3907168a9