Additional refactoring to allow adding more tests
Additional refactoring is done in order to make classes testable. Now,
instead of passing the parameters to establish a connection to the whole
chain of classes, dependency injection is used to simplify the class
dependencies.

Given the injection stack is limited when doing the site initialization,
circumvent it by calling directly the class responsible for establishing
the connection.

Change-Id: I52f22e98299852833e9e833b9879284a11c920b0