Add tests to JiraItsFacade
Passing the JiraClient dependency as parameter to the constructor
(instead of creating it inside the class) helps making this class
testable. In the same sense, the Jira configuration can be injected now
into the JiraClient; the previous constructor is kept, though, to allow
the init step to test connectivity. This will be modified in a follow up
change to avoid passing parameters along the whole chain of calls.

Change-Id: Iee6803aded52fff70e079630bed421748ee27d68