Update patch set 2
Patch Set 2:

Yes, you are right. It was not working because of the previous change.

This was caused by the method URL(context,spec) method used in JiraRestApi -
private HttpURLConnection prepHttpConnection(String spec, boolean isPostRequest)

When trying to create a new URL by providing already existing URL as the first parameter in URL(context,spec) method, the path of the URL was getting replaced by the "spec" part which caused the bug. 

I fixed the bug by postponing the conversion of string to URL and is only done once in prepHttpConnection(..) method now.

Patch-set: 2