Fix for the issue not found bug[1]
The bug caused by the change[2].The method URL(context,spec) in
JiraRestApi caused the issue. If the spec begins with a forward
slash "/", it replaces the path provided in the already passed URL
context as a first parameter. It is fixed by appending the spec to the
URL string and converting it directly into the URL.

[1] https://bugs.chromium.org/p/gerrit/issues/detail?id=8263
[2] https://gerrit-review.googlesource.com/c/plugins/its-jira/+/144970

Change-Id: I3ff701b25d9c69a4ebed88e3ff07f6390ea177a9