Fix Jira issue not found
After I9a0183b58, plugin was able to connect to a Jira server but failed
to find or annotate issues. The method URL(context,spec) in JiraRestApi
caused the issue: If the spec begins with a forward slash, it replaces
the path provided in the passed URL context as a first parameter.
This is fixed by appending the spec to the URL string and converting it
directly into the URL.

Bug: Issue 8263
Change-Id: I3ff701b25d9c69a4ebed88e3ff07f6390ea177a9