Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> (4 comments)

You are absolutely right. I was confused due to two different types of configuration styles documented in the individual plugin's documentation and the general Gerrit documentation.
The below documentation got me confused:
https://gerrit-review.googlesource.com/Documentation/dev-plugins.html#simple-configuration

Reading this implies that all plugins should be configured with "plugin" token as the section in gerrit.config. But reading the documentation of every its-@plugin@, the documentation seems inconsistent with the Gerrit documentation but consistent amongst the its-family. So, I think this change can be abandoned and a change in the Gerrit documentation is required to make it consistent with the plugins. Waiting for your suggestion.

Patch-set: 11
Reviewer: Inderjot Kaur Ratol <1055597@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0