Update patch set 22
Patch Set 22:

> Patch Set 22:
> 
> Shall we put this into WIP until the encryption solution has been defined?
> Or, better, why don't we merge this without the encryption and delegate that piece to a follow-up change?

The encryption change does not seem to be going well so far, so that could delay more this one, which is the main reason we started all this. I see two options: either we go with no encryption (which is not good) or we go back to a plugin-only implementation (as proposed initially) and once the core change is in, we move to the new way. Other (better) options?

Patch-set: 22
Reviewer: Hector Oswaldo Caballero <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>