Update patch set 24
Patch Set 24:

> Patch Set 24:
> 
> @Hector have you decided on how to proceed with this change? What about my proposal to declare it "experimental" in the documentation and de-scope the encryption for now?
> 
> It would be nice to have it "landed" to the master branch ASAP.

IIUC, your proposal is to merge this one without encryption? In that case, the passwords will be plain text in the project configuration. Is this acceptable? For the encryption, we are working on a proposal so that the community agrees before starting the implementation.

Patch-set: 24
Reviewer: Hector Oswaldo Caballero <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>