Update patch set 35
Patch Set 35:

> Patch Set 35:
> 
> > Patch Set 35: Verified-1
> > 
> > Tests are not working on master. Let me check against Gerrit v2.16: if they work, I'll create a stable-2.16 branch and move this change to that one.
> 
> Do you mean your own manual tests, as previously mentioned? Because the automated plugin ones are still passing, at least for me locally.

The automated tests are passing on stable-2.16, but they do not even build on master :-(

> 
> > Then, once merged, I'll merge it back to master and fix the remaining issues.
> > 
> > @Marco what do you think about the plan?
> 
> Sure. Sorry for the delay in replying, was stuck reinstalling my computer for the last few days (new corporate one). I suppose you propose 2.16 because you no longer deploy 2.15?

Because v2.16 is the latest active version of Gerrit, so people may expect new features in its-jira also.

> We are still deploying 2.14 with a corresponding version of this change in.

Yes, I remember that. We could try to see if this merges cleanly on v2.15 also, so that you guys can get rid of the fork as soon as you upgrade to that version. WDYT?

Patch-set: 35
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>