Update patch set 35
Patch Set 35:

> The automated tests are passing on stable-2.16, but they do not even build on master :-(

When I run the plugin tests locally on master (gerrit tree), they all built and passed for me. But maybe you are referring to a CI run of those, as opposed to a typical bazel plugin one?

> > We are still deploying 2.14 with a corresponding version of this change in.
> 
> Yes, I remember that. We could try to see if this merges cleanly on v2.15 also, so that you guys can get rid of the fork as soon as you upgrade to that version. WDYT?

We will need to revisit our its forks for an upcoming 2.15 deployment, yes. I agree with the idea of aiming for a 2.15 version of this change first then merge-up again. Would not lose the work done in this very master/2.16 change for when the latter unfolds.

Patch-set: 35
Reviewer: Gerrit User 1012732 <1012732@173816e5-2b9a-37c3-8a2e-48639d4f1153>