Fix its-jira init step
Since the introduction of the multi-jira servers in v2.16, the way
the Jira REST-API URL path is composed has changed. The InitJira step
has not been adapted to the new URL syntax and thus was not working
properly.

Remove the trailing slash in the Jira init step when requesting
for the server info and the associated Jira version.

Also the rendering of the version on the init output was incorrect
resulting in unexpanded {} delimiters.

Bug: Issue 10639
Change-Id: I3dd6d520efd7f0e58251e558365e3b5a17d1983f