Reintroduce Jira URL normalization
Fix a regression introduced in v2.16 where the existing configurations
were broken because the code was assuming that all URLs are ending with
'/'.

Apparently the condition was covered by tests, but the normalization of
the URL, instead of being present in the plugin code, was done in the tests
making them ineffective.

Bug: Issue 11308
Change-Id: Ie1d25e7207e3ef862e848d94d14f224470d629e0